# docs/openapi.yaml
openapi: 3.0.3
info:
  title: WTR Ghana Water Platform API
  version: 1.0.0
  description: Multi-tenant community water utility management API.
paths:
...
  /api/communities/:
    get:
      operationId: communities_list
      parameters:
      - in: query
        name: district
        schema:
          type: string
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: region
        schema:
          type: string
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: service_status
        schema:
          type: string
          enum:
          - ACTIVE
          - INACTIVE
          - PENDING
          - SUSPENDED
        description: |-
          * `PENDING` - Pending approval
          * `ACTIVE` - Active
          * `SUSPENDED` - Suspended
          * `INACTIVE` - Inactive
      - in: query
        name: subscription_plan
        schema:
          type: integer
      - in: query
        name: water_system_type
        schema:
          type: string
          enum:
          - BOREHOLE
          - HANDPUMP
          - MIXED
          - PIPED
          - SMALL_TOWN
        description: |-
          * `PIPED` - Piped scheme
          * `BOREHOLE` - Mechanised borehole
          * `SMALL_TOWN` - Small-town water system
          * `HANDPUMP` - Hand-pump
          * `MIXED` - Mixed
      tags:
      - communities
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCommunityList'
          description: ''
    post:
      operationId: communities_create
      tags:
      - communities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Community'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Community'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Community'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
          description: ''
  /api/communities/{id}/:
    get:
      operationId: communities_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this community.
        required: true
      tags:
      - communities
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
          description: ''
    put:
      operationId: communities_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this community.
        required: true
      tags:
      - communities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Community'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Community'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Community'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
          description: ''
    patch:
      operationId: communities_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this community.
        required: true
      tags:
      - communities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCommunity'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCommunity'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCommunity'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
          description: ''
    delete:
      operationId: communities_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this community.
        required: true
      tags:
      - communities
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
...
  /api/communities/{id}/features/:
    get:
      operationId: communities_features_retrieve
      description: |-
        System settings: which feature modules each role in this community may use, plus
        community-wide switches (customer portal, online payments…). Community admin only.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this community.
        required: true
      tags:
      - communities
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
          description: ''
    patch:
      operationId: communities_features_partial_update
      description: |-
        System settings: which feature modules each role in this community may use, plus
        community-wide switches (customer portal, online payments…). Community admin only.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this community.
        required: true
      tags:
      - communities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCommunity'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCommunity'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCommunity'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
          description: ''
  /api/communities/{id}/settings/:
    get:
      operationId: communities_settings_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this community.
        required: true
      tags:
      - communities
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
          description: ''
    patch:
      operationId: communities_settings_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this community.
        required: true
      tags:
      - communities
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCommunity'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCommunity'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCommunity'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Community'
          description: ''
...
components:
  schemas:
...
    Community:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        admin_name:
          type: string
          readOnly: true
        admin_email:
          type: string
          readOnly: true
        plan_name:
          type: string
          readOnly: true
        customers_total:
          type: string
          readOnly: true
        active_meters:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        code:
          type: string
          readOnly: true
        name:
          type: string
          maxLength: 150
        region:
          type: string
          maxLength: 80
        district:
          type: string
          maxLength: 100
        municipality:
          type: string
          maxLength: 100
        town:
          type: string
          maxLength: 100
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        boundary:
          nullable: true
          description: GeoJSON polygon of the community boundary
        contact_name:
          type: string
          maxLength: 120
        contact_phone:
          type: string
          maxLength: 30
        contact_email:
          oneOf:
          - type: string
            format: email
            maxLength: 254
          - type: string
            maxLength: 0
        water_source:
          type: string
          maxLength: 120
        water_system_type:
          $ref: '#/components/schemas/WaterSystemTypeEnum'
        households_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        meters_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        service_status:
          $ref: '#/components/schemas/ServiceStatusEnum'
        approved_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        registration_date:
          type: string
          format: date
          readOnly: true
        notes:
          type: string
        public_report_token:
          type: string
          maxLength: 32
        admin:
          type: integer
          nullable: true
        approved_by:
          type: integer
          readOnly: true
          nullable: true
        subscription_plan:
          type: integer
          nullable: true
      required:
      - active_meters
      - admin_email
      - admin_name
      - approved_at
      - approved_by
      - code
      - created_at
      - customers_total
      - district
      - id
      - name
      - plan_name
      - region
      - registration_date
      - updated_at
...
    PaginatedCommunityList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Community'
...
    PatchedCommunity:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        admin_name:
          type: string
          readOnly: true
        admin_email:
          type: string
          readOnly: true
        plan_name:
          type: string
          readOnly: true
        customers_total:
          type: string
          readOnly: true
        active_meters:
          type: string
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        code:
          type: string
          readOnly: true
        name:
          type: string
          maxLength: 150
        region:
          type: string
          maxLength: 80
        district:
          type: string
          maxLength: 100
        municipality:
          type: string
          maxLength: 100
        town:
          type: string
          maxLength: 100
        latitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        longitude:
          type: string
          format: decimal
          pattern: ^-?\d{0,3}(?:\.\d{0,6})?$
          nullable: true
        boundary:
          nullable: true
          description: GeoJSON polygon of the community boundary
        contact_name:
          type: string
          maxLength: 120
        contact_phone:
          type: string
          maxLength: 30
        contact_email:
          oneOf:
          - type: string
            format: email
            maxLength: 254
          - type: string
            maxLength: 0
        water_source:
          type: string
          maxLength: 120
        water_system_type:
          $ref: '#/components/schemas/WaterSystemTypeEnum'
        households_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        meters_count:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        service_status:
          $ref: '#/components/schemas/ServiceStatusEnum'
        approved_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        registration_date:
          type: string
          format: date
          readOnly: true
        notes:
          type: string
        public_report_token:
          type: string
          maxLength: 32
        admin:
          type: integer
          nullable: true
        approved_by:
          type: integer
          readOnly: true
          nullable: true
        subscription_plan:
          type: integer
          nullable: true
...
    ServiceStatusEnum:
      enum:
      - PENDING
      - ACTIVE
      - SUSPENDED
      - INACTIVE
      type: string
      description: |-
        * `PENDING` - Pending approval
        * `ACTIVE` - Active
        * `SUSPENDED` - Suspended
        * `INACTIVE` - Inactive
...
    WaterSystemTypeEnum:
      enum:
      - PIPED
      - BOREHOLE
      - SMALL_TOWN
      - HANDPUMP
      - MIXED
      type: string
      description: |-
        * `PIPED` - Piped scheme
        * `BOREHOLE` - Mechanised borehole
        * `SMALL_TOWN` - Small-town water system
        * `HANDPUMP` - Hand-pump
        * `MIXED` - Mixed
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT